+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- sample app enhancements
+
+### Deprecated
+
+- python 3.7 support will be removed on 2024-06-27
+
+### Removed
+
+- py3.7: no longer run checks
+
 ## [0.1.25] - 2023-06-07
 
 ### Changed